Unit Test for Update End point
Added update end point Unit test
