+++ TechLibraryExercise-master/TechLibrary.Web/TechLibrary.Web.njsproj
@@ -60,8 +60,8 @@
     <Content Include="src\main.js" />
   </ItemGroup>
   <ItemGroup>
-    <TypeScriptCompile Include="src\components\AddBook\addBook.ts" />
-    <TypeScriptCompile Include="src\components\Book\book.ts" />
+    <TypeScriptCompile Include="src\components\AddBook\addBook.js" />
+    <TypeScriptCompile Include="src\components\Book\book.js" />
   </ItemGroup>
   <PropertyGroup>
     <PostBuildEvent>
