+++ TechLibraryExercise-master/TechLibrary.Web/src/components/AddBook/addBook.js
@@ -1,9 +1,8 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-var axios_1 = require("axios");
-exports.default = {
+import axios from 'axios';
+
+export default {
     name: 'AddBook',
-    data: function () {
+    data() {
         return {
             form: {
                 title: '',
@@ -15,33 +14,30 @@ exports.default = {
             },
             show: true,
             success: false
-        };
+        }
     },
     methods: {
-        onSubmit: function (evt) {
-            var _this = this;
-            evt.preventDefault();
-            alert(JSON.stringify(this.form));
-            axios_1.default.post("https://localhost:5001/books", this.form)
-                .then(function (response) {
-                _this.success = response.data;
-                evt.target.reset();
-            });
+        onSubmit(evt) {
+            evt.preventDefault()
+            alert(JSON.stringify(this.form))
+            axios.post(`https://localhost:5001/books`, this.form)
+                .then(response => {
+                    this.success = response.data;
+                    evt.target.reset();
+                });
         },
-        onReset: function (evt) {
-            var _this = this;
-            evt.preventDefault();
-            this.form.title = '';
-            this.form.isbn = '';
-            this.form.publishedDate = '';
-            this.form.shortDescr = '';
-            this.form.longDescr = '';
-            this.form.thumbnailurl = '';
-            this.show = false;
-            this.$nextTick(function () {
-                _this.show = true;
-            });
+        onReset(evt) {
+            evt.preventDefault()
+            this.form.title = ''
+            this.form.isbn = ''
+            this.form.publishedDate = ''
+            this.form.shortDescr = ''
+            this.form.longDescr = ''
+            this.form.thumbnailurl = ''
+            this.show = false
+            this.$nextTick(() => {
+                this.show = true
+            })
         }
     }
-};
-//# sourceMappingURL=addBook.js.map
+}
